Result: remove 'resultReallyIn' field. Instead, 'ResultSource' type is introduced.
This is because we are gonna aggregate 'Result's and its sources, so
it's more convenient to have "queried source" and "real source"
combined.

Also, implementation of Exec is changed so it now ALWAYS sets
'resultSourceReal' field if it loads the BuildPlan successfully. I think
this makes much more sense. Due to this change, some tests fail now.